New info style and some cleaning up of solvers
So far so good: point jet appears to give correct
results and also the kolmogorov forcing seems to be
fine. Extensions from NL to GQL and GCE2 required.

Maybe there is a possibility to be simpler than the
current multiple dispatch framework, but we will
stick to this framework for now, which will lead to
9 methods separately defined for solver->forcing
combinations.

An important part is going to be the time dependent
forcing, which might affect the equations method.