Use global dist defined in noise.jl for NL/GQL and GCE2
Noise defines the S&Y12 isotropic distribution. For NL, GQL
and the low mode GCE2, stochastic forcing occurs. For the field
bilinear, the covariance is added to the tendency directly
via the .x[2] component of array partitions.

A further change in ICs has been made to ensure GCE2 uses a GQL
type IC after Random seeding. The GQL/NL ic for random values is
changed to use a random phase with amplitude prescribed.

Coefficients for forcing are pre-computed like all coefficients.
These must now be added to other solver routines also because F
is part of p, but this should allow a globalisation of the solver
in due course, a move towards unifying the various dispatches or
at least vast sections of dispatches.

CE2 and QL are off by a small factor which needs time averaging.